# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions/{slug}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get question detail by slug (increments view count)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Question detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /admin/sponsor-inquiries:
    get:
      operationId: listSponsorInquiriesAdmin
      tags: [admin]
      summary: List sponsor inquiries (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, closed]
      responses:
        "200":
          description: Sponsor inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SponsorInquiry"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /admin/sponsor-inquiries/{id}:
    patch:
      operationId: updateSponsorInquiryAdmin
      tags: [admin]
      summary: Update a sponsor inquiry's follow-up status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorInquiryStatusUpdate"
      responses:
        "200":
          description: Updated inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorInquiry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: Admin dashboard stats
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /admin/flags/{id}/resolve:
    post:
      operationId: resolveFlag
      tags: [admin]
      summary: Resolve a flag (dismiss or remove content)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagResolution"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /admin/users:
    get:
      operationId: listUsersAdmin
      tags: [admin]
      summary: List users for moderation
      parameters:
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}:
    patch:
      operationId: updateUserAdmin
      tags: [admin]
      summary: Update user role / suspension
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAdminUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/questions/{id}:
    delete:
      operationId: deleteQuestionAdmin
      tags: [admin]
      summary: Remove a question
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    patch:
      operationId: featureQuestionAdmin
      tags: [admin]
      summary: Pin/unpin (feature) a question
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionFeatureUpdate"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /admin/answers/{id}:
    delete:
      operationId: deleteAnswerAdmin
      tags: [admin]
      summary: Remove an answer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /admin/tags:
    post:
      operationId: createTagAdmin
      tags: [admin]
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreate"
      responses:
        "201":
          description: Created tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagInfo"
  /admin/categories:
    post:
      operationId: createCategoryAdmin
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryInfo"
  /admin/categories/{id}:
    patch:
      operationId: updateCategoryAdmin
      tags: [admin]
      summary: Edit a category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryInfo"
    delete:
      operationId: deleteCategoryAdmin
      tags: [admin]
      summary: Delete a category (questions keep working, become uncategorized)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /admin/tags/{id}:
    patch:
      operationId: updateTagAdmin
      tags: [admin]
      summary: Edit a tag
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpdate"
      responses:
        "200":
          description: Updated tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagInfo"
    delete:
      operationId: deleteTagAdmin
      tags: [admin]
      summary: Delete a tag
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message: { type: string }
      required: [message]
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        reputation: { type: integer }
        role:
          type: string
          enum: [member, moderator, admin, platform_owner]
        plan:
          type: string
          enum: [free, pro]
        isSuspended: { type: boolean }
        questionCount: { type: integer }
        answerCount: { type: integer }
        acceptedAnswerCount: { type: integer }
        createdAt: { type: string }
      required:
        [id, username, displayName, reputation, role, plan, isSuspended, questionCount, answerCount, acceptedAnswerCount, createdAt]
...
    UserAdminUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [member, moderator, admin]
        isSuspended: { type: boolean }
...
    Question:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        body: { type: string }
        language: { type: string }
        author:
          $ref: "#/components/schemas/User"
        tags:
          type: array
          items: { type: string }
        category:
          oneOf:
            - $ref: "#/components/schemas/CategoryRef"
            - type: "null"
        score: { type: integer }
        answerCount: { type: integer }
        viewCount: { type: integer }
        hasAcceptedAnswer: { type: boolean }
        isFeatured: { type: boolean }
        myVote: { type: ["integer", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        [id, slug, title, body, language, author, tags, score, answerCount, viewCount, hasAcceptedAnswer, isFeatured, createdAt, updatedAt]
    QuestionDetail:
      type: object
      properties:
        question:
          $ref: "#/components/schemas/Question"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      required: [question, answers, comments]
...
    QuestionFeatureUpdate:
      type: object
      properties:
        isFeatured: { type: boolean }
      required: [isFeatured]
    Answer:
      type: object
      properties:
        id: { type: integer }
        questionId: { type: integer }
        body: { type: string }
        language: { type: string }
        author:
          $ref: "#/components/schemas/User"
        score: { type: integer }
        isAccepted: { type: boolean }
        myVote: { type: ["integer", "null"] }
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt: { type: string }
      required: [id, questionId, body, language, author, score, isAccepted, comments, createdAt]
...
    Comment:
      type: object
      properties:
        id: { type: integer }
        parentType:
          type: string
          enum: [question, answer]
        parentId: { type: integer }
        body: { type: string }
        author:
          $ref: "#/components/schemas/User"
        createdAt: { type: string }
      required: [id, parentType, parentId, body, author, createdAt]
...
    TagInfo:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        questionCount: { type: integer }
      required: [id, slug, name, questionCount]
    TagUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 50 }
        description: { type: string, maxLength: 500 }
    TagCreate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 50 }
        description: { type: string, maxLength: 500 }
      required: [name]
    CategoryRef:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
      required: [slug, name]
    CategoryInfo:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        questionCount: { type: integer }
      required: [id, slug, name, questionCount]
    CategoryCreate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        description: { type: string, maxLength: 500 }
      required: [name]
    CategoryUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        description: { type: string, maxLength: 500 }
...
    FlagResolution:
      type: object
      properties:
        action:
          type: string
          enum: [dismiss, remove_content]
      required: [action]
...
    SponsorInquiry:
      type: object
      properties:
        id: { type: integer }
        company: { type: string }
        contactName: { type: string }
        email: { type: string }
        budgetRange:
          type: string
          enum: [under_1k, 1k_5k, 5k_20k, over_20k, undecided]
        message: { type: string }
        status:
          type: string
          enum: [new, contacted, closed]
        createdAt: { type: string }
      required: [id, company, contactName, email, budgetRange, message, status, createdAt]
    SponsorInquiryStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, closed]
      required: [status]
...
    AdminOverview:
      type: object
      properties:
        questionCount: { type: integer }
        answerCount: { type: integer }
        userCount: { type: integer }
        openFlagCount: { type: integer }
        questionsToday: { type: integer }
        answersToday: { type: integer }
        newUsersToday: { type: integer }
      required: [questionCount, answerCount, userCount, openFlagCount, questionsToday, answersToday, newUsersToday]
...